pyproject.toml
[project]
name = "kinecho-main"
version = "0.0.0.2"
description = "AI companion Kinecho"
authors = [
    {name = "LilyAviarn"}
]
license = {text = "None"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair (>=5.5.0,<6.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "configparser (>=7.2.0,<8.0.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "google-cloud-translate (>=3.20.2,<4.0.0)",
    "kafka-python (>=2.2.10,<3.0.0)",
    "mss (>=10.0.0,<11.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "pydeck (>=0.9.1,<0.10.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyttsx3 (>=2.98,<3.0)",
    "requests (>=2.32.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "speechrecognition (>=3.14.3,<4.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "watchdog (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.uv]
package = false
